Fix map gen and add grid movement
Took 3 hours 27 minutes

+++ src/entities/common.rs
@@ -1,3 +1,3 @@
 pub struct Health {
-    pub hp: f32,
+    pub hp: f64,
 }
